+++ packages/slidev-theme-eloc/setup/windicss.ts
@@ -4,7 +4,7 @@ import { defineWindiSetup } from '@slidev/types'
 export default defineWindiSetup(() => ({
   extract: {
     include: [
-      resolve(__dirname, '**/*.vue'),
+      resolve(__dirname, '../**/*.{vue,ts}'),
     ],
     exclude: [
       resolve(__dirname, 'node_modules'),